+++ feature/main/src/main/java/com/ku_stacks/ku_ring/main/MainActivity.kt
@@ -1,6 +1,5 @@
 package com.ku_stacks.ku_ring.main
 
-import android.Manifest
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
@@ -8,31 +7,21 @@ import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.Settings
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
 import androidx.core.content.IntentCompat
 import androidx.navigation.NavHostController
 import androidx.lifecycle.compose.LifecycleResumeEffect
 import androidx.navigation.compose.rememberNavController
-import com.ku_stacks.ku_ring.designsystem.components.KuringAlertDialog
 import com.ku_stacks.ku_ring.designsystem.kuringtheme.KuringTheme
 import com.ku_stacks.ku_ring.domain.WebViewNotice
 import com.ku_stacks.ku_ring.navigation.KuringNavigator
 import com.ku_stacks.ku_ring.navigation.MainScreenRoute
 import com.ku_stacks.ku_ring.preferences.PreferenceUtil
+import com.ku_stacks.ku_ring.ui_util.KuringNavigator
 import com.ku_stacks.ku_ring.util.checkHasNotificationPermission
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
@@ -81,48 +70,12 @@ class MainActivity : AppCompatActivity() {
             navController = rememberNavController()
             val startDestination = intent.parseMainScreenRoute() ?: MainScreenRoute.Notice
             KuringTheme {
-                val context = LocalContext.current
-                var isOpenSettingsDialogVisible by rememberSaveable { mutableStateOf(false) }
-                var isAppSettingsOpened by rememberSaveable { mutableStateOf(false) }
-
-                // 앱 설정 이동 후에 다시 확인
-                LifecycleResumeEffect(Unit) {
-                    if (context.checkHasNotificationPermission()
-                        && isAppSettingsOpened
-                    ) {
-                        pref.resetNotificationPermissionDialogCount()
-                        isAppSettingsOpened = false
-                    }
-                    onPauseOrDispose {}
-                }
-
-                RequestNotificationPermission(
-                    onGranted = {
-                        pref.resetNotificationPermissionDialogCount()
-                    },
-                    onDenied = {
-                        if (pref.canShowNotificationPermissionDialog()) {
-                            isOpenSettingsDialogVisible = true
-                        }
-                    },
+                NotificationPermissionHandler(
+                    onPermissionGranted = { pref.resetNotificationPermissionDialogCount() },
+                    canShowSettingsDialog = { pref.canShowNotificationPermissionDialog() },
+                    onSettingsDialogShown = { pref.notificationPermissionDialogCount++ },
+                    openAppNotificationSettings = { launchAppNotificationSettings() }
                 )
-
-                if (isOpenSettingsDialogVisible) {
-                    KuringAlertDialog(
-                        text = stringResource(R.string.notification_permission_dialog_text),
-                        confirmText = stringResource(R.string.notification_permission_dialog_confirm),
-                        onConfirm = {
-                            isAppSettingsOpened = true
-                            isOpenSettingsDialogVisible = false
-                            pref.notificationPermissionDialogCount++
-                            launchAppSettings()
-                        },
-                        onCancel = {
-                            isOpenSettingsDialogVisible = false
-                            pref.notificationPermissionDialogCount++
-                        },
-                    )
-                }
 
                 val navController = rememberNavController()
                 MainScreen(
@@ -136,26 +89,11 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    @Composable
-    private fun RequestNotificationPermission(
-        onGranted: () -> Unit = {},
-        onDenied: () -> Unit = {},
-    ) {
-        val context = LocalContext.current
-        if (!context.checkHasNotificationPermission()) {
-            val launcher = rememberLauncherForActivityResult(
-                contract = ActivityResultContracts.RequestPermission(),
-            ) { isGranted ->
-                if (isGranted) onGranted() else onDenied()
-            }
-
-            LaunchedEffect(Unit) {
-                launcher.launch(Manifest.permission.POST_NOTIFICATIONS)
-            }
-        }
+    private fun navToNoticeActivity(webViewNotice: WebViewNotice) {
+        navigator.navigateToNoticeWeb(this, webViewNotice)
     }
 
-    private fun launchAppSettings() {
+    private fun launchAppNotificationSettings() {
         val intent = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
             Intent(Settings.ACTION_APP_NOTIFICATION_SETTINGS).apply {
                 putExtra(Settings.EXTRA_APP_PACKAGE, packageName)
